bot/requirements.txt
requests==2.31.0
numpy==1.26.4
pipecat-ai==0.1.38
websockets==11.0.3
